Fixups to work with new mana sycophant
When we merged sycophant code into mana, I made the state dir
configurable and changed some of the filenames. The config option still
needs to be done here, but if you set it to /tmp in mana it'll work.

Also enabled libnl32 since more mana users seem to use that.

Corrected the spelling of gcp's name.